put helper function in a different file
