+++ src/components/Popover/index.jsx
@@ -3,18 +3,7 @@ import PropTypes from 'prop-types';
 import Tippy from '@tippyjs/react';
 import FDS from 'lib/dictionary/js/styleConstants';
 import hideOnEscFunc from './hideOnEsc';
-
-const triggerHelper = (trigger, isOpen) => {
-  // https://atomiks.github.io/tippyjs/v6/all-props/#trigger
-  // We'd like hover to semantically map to mouseenter + focus
-  const triggerMapping = {
-    click: 'click',
-    hover: 'mouseenter focus',
-  };
-  /* Tippy will throw a warning if you attempt to pass a trigger when
-     isOpen is passed. This will avoid showing that warning. */
-  return isOpen !== undefined ? undefined : triggerMapping[trigger];
-};
+import triggerHelper from './triggerHelper';
 
 const Popover = ({
   children,
